refactor(ui): split realtime logging models
Refs #2496
